move clock into SpecFlow project
